LPS-58144 Moves from com.liferay.portal.kernel.portlet.configuration.* to com.liferay.portal.kernel.portlet.configuration.icon.
