+++ modules/addons/portlet-configuration-icon/portlet-configuration-icon-export-import/src/com/liferay/portlet/configuration/icon/exportimport/ExportImportPortletConfigurationIcon.java
@@ -15,7 +15,7 @@
 package com.liferay.portlet.configuration.icon.exportimport;
 
 import com.liferay.portal.kernel.language.LanguageUtil;
-import com.liferay.portal.kernel.portlet.configuration.BasePortletConfigurationIcon;
+import com.liferay.portal.kernel.portlet.configuration.icon.BasePortletConfigurationIcon;
 import com.liferay.portal.kernel.util.HtmlUtil;
 import com.liferay.portal.kernel.util.StringBundler;
 import com.liferay.portal.theme.PortletDisplay;
