+++ modules/addons/portlet-configuration-icon/portlet-configuration-icon-staging/src/com/liferay/portlet/configuration/icon/staging/StagingPortletConfigurationIcon.java
@@ -15,7 +15,7 @@
 package com.liferay.portlet.configuration.icon.staging;
 
 import com.liferay.portal.kernel.language.LanguageUtil;
-import com.liferay.portal.kernel.portlet.configuration.BasePortletConfigurationIcon;
+import com.liferay.portal.kernel.portlet.configuration.icon.BasePortletConfigurationIcon;
 import com.liferay.portal.kernel.util.HtmlUtil;
 import com.liferay.portal.kernel.util.StringBundler;
 import com.liferay.portal.theme.PortletDisplay;
