+++ modules/addons/portlet-configuration-icon/portlet-configuration-icon-widget/src/com/liferay/portlet/configuration/icon/widget/WidgetPortletConfigurationIcon.java
@@ -17,7 +17,7 @@ package com.liferay.portlet.configuration.icon.widget;
 import com.liferay.portal.kernel.exception.PortalException;
 import com.liferay.portal.kernel.log.Log;
 import com.liferay.portal.kernel.log.LogFactoryUtil;
-import com.liferay.portal.kernel.portlet.configuration.BasePortletConfigurationIcon;
+import com.liferay.portal.kernel.portlet.configuration.icon.BasePortletConfigurationIcon;
 import com.liferay.portal.kernel.util.GetterUtil;
 import com.liferay.portal.kernel.util.StringBundler;
 import com.liferay.portal.kernel.util.StringPool;
