+++ modules/addons/portlet-configuration-icon/portlet-configuration-icon-widget/src/com/liferay/portlet/configuration/icon/widget/WidgetPortletConfigurationIconFactory.java
@@ -14,9 +14,9 @@
 
 package com.liferay.portlet.configuration.icon.widget;
 
-import com.liferay.portal.kernel.portlet.configuration.BasePortletConfigurationIconFactory;
-import com.liferay.portal.kernel.portlet.configuration.PortletConfigurationIcon;
-import com.liferay.portal.kernel.portlet.configuration.PortletConfigurationIconFactory;
+import com.liferay.portal.kernel.portlet.configuration.icon.BasePortletConfigurationIconFactory;
+import com.liferay.portal.kernel.portlet.configuration.icon.PortletConfigurationIcon;
+import com.liferay.portal.kernel.portlet.configuration.icon.PortletConfigurationIconFactory;
 
 import javax.servlet.http.HttpServletRequest;
