+++ modules/apps/journal/journal-web/src/META-INF/resources/configuration/icon/structures.jsp
