+++ modules/apps/journal/journal-web/src/META-INF/resources/configuration/icon/templates.jsp
