+++ modules/apps/journal/journal-web/src/com/liferay/journal/web/portlet/configuration/icon/FeedsPortletConfigurationIconFactory.java
@@ -12,11 +12,11 @@
  * details.
  */
 
-package com.liferay.journal.web.portlet.configuration;
+package com.liferay.journal.web.portlet.configuration.icon;
 
-import com.liferay.portal.kernel.portlet.configuration.BaseJSPPortletConfigurationIconFactory;
-import com.liferay.portal.kernel.portlet.configuration.PortletConfigurationIcon;
-import com.liferay.portal.kernel.portlet.configuration.PortletConfigurationIconFactory;
+import com.liferay.portal.kernel.portlet.configuration.icon.BaseJSPPortletConfigurationIconFactory;
+import com.liferay.portal.kernel.portlet.configuration.icon.PortletConfigurationIcon;
+import com.liferay.portal.kernel.portlet.configuration.icon.PortletConfigurationIconFactory;
 
 import javax.servlet.ServletContext;
 import javax.servlet.http.HttpServletRequest;
@@ -28,7 +28,7 @@ import org.osgi.service.component.annotations.Reference;
  * @author Eudaldo Alonso
  */
 @Component(immediate = true, service = PortletConfigurationIconFactory.class)
-public class StructuresPortletConfigurationIconFactory
+public class FeedsPortletConfigurationIconFactory
 	extends BaseJSPPortletConfigurationIconFactory {
 
 	@Override
@@ -38,12 +38,12 @@ public class StructuresPortletConfigurationIconFactory
 
 	@Override
 	public String getJspPath() {
-		return "/configuration/structures.jsp";
+		return "/configuration/icon/feeds.jsp";
 	}
 
 	@Override
 	public double getWeight() {
-		return 102;
+		return 100.0;
 	}
 
 	@Override
