+++ modules/apps/journal/journal-web/src/com/liferay/journal/web/portlet/configuration/icon/TemplatesPortletConfigurationIconFactory.java
@@ -12,11 +12,11 @@
  * details.
  */
 
-package com.liferay.journal.web.portlet.configuration;
+package com.liferay.journal.web.portlet.configuration.icon;
 
-import com.liferay.portal.kernel.portlet.configuration.BaseJSPPortletConfigurationIconFactory;
-import com.liferay.portal.kernel.portlet.configuration.PortletConfigurationIcon;
-import com.liferay.portal.kernel.portlet.configuration.PortletConfigurationIconFactory;
+import com.liferay.portal.kernel.portlet.configuration.icon.BaseJSPPortletConfigurationIconFactory;
+import com.liferay.portal.kernel.portlet.configuration.icon.PortletConfigurationIcon;
+import com.liferay.portal.kernel.portlet.configuration.icon.PortletConfigurationIconFactory;
 
 import javax.servlet.ServletContext;
 import javax.servlet.http.HttpServletRequest;
@@ -28,7 +28,7 @@ import org.osgi.service.component.annotations.Reference;
  * @author Eudaldo Alonso
  */
 @Component(immediate = true, service = PortletConfigurationIconFactory.class)
-public class FeedsPortletConfigurationIconFactory
+public class TemplatesPortletConfigurationIconFactory
 	extends BaseJSPPortletConfigurationIconFactory {
 
 	@Override
@@ -38,12 +38,12 @@ public class FeedsPortletConfigurationIconFactory
 
 	@Override
 	public String getJspPath() {
-		return "/configuration/feeds.jsp";
+		return "/configuration/icon/templates.jsp";
 	}
 
 	@Override
 	public double getWeight() {
-		return 100.0;
+		return 101;
 	}
 
 	@Override
