+++ portal-service/src/com/liferay/portal/kernel/portlet/configuration/icon/packageinfo
